feat: 메인 페이지 카드 타입 정의
